Refactor: /service 하위 파일 중 lib, component 분리
