+++ src/app/(main)/activity/[semesterName]/[activityId]/boards/[boardId]/_components/BoardHero.tsx
@@ -11,7 +11,7 @@ import {
 } from '@/components/feature'
 import { Button, Separator, useToast } from '@/components/ui'
 import { DATA_ERROR_MESSAGES } from '@/constant/errorMessage'
-import { queryClient } from '@/service/components/ReactQueryClientProvider'
+import { queryClient } from '@/lib/query-client'
 import { boardDetailQuery } from '@/service/data/boards'
 import { deleteBoardAction } from '@/service/server/board/delete-board'
 import { useMyInfoStore } from '@/store/myInfo'
