+++ src/app/(main)/admin/member/_components/approve-table/Button.tsx
@@ -4,7 +4,7 @@ import { useAction } from 'next-safe-action/hooks'
 import { useRouter } from 'next/navigation'
 
 import { Button, ToastAction, useToast } from '@/components/ui'
-import { queryClient } from '@/service/components/ReactQueryClientProvider'
+import { queryClient } from '@/lib/query-client'
 import {
   approveMemberAction,
   rejectMemberAction,
