+++ src/app/(main)/notice/create-post/_components/Form.tsx
@@ -9,8 +9,8 @@ import { usePathname, useRouter } from 'next/navigation'
 
 import { CreatePostForm } from '@/components/feature'
 import { useToast } from '@/components/ui'
+import { queryClient } from '@/lib/query-client'
 import { CreateNoticePost, CreateNoticePostSchema } from '@/schema/post'
-import { queryClient } from '@/service/components/ReactQueryClientProvider'
 import { createNoticePostAction } from '@/service/server/post/create-post'
 
 export const CreateNoticePostForm = () => {
