+++ src/app/(main)/notice/create-post/layout.tsx
@@ -4,8 +4,7 @@ import { ReactNode } from 'react'
 
 import { useRouter } from 'next/navigation'
 
-import { MainErrorFallback } from '@/components/common'
-import ErrorBoundary from '@/service/components/error-boundary'
+import { ErrorBoundary, MainErrorFallback } from '@/components/common'
 
 const CreateNoticePostLayout = ({ children }: { children: ReactNode }) => {
   const router = useRouter()
