+++ src/components/common/error-boundary/index.ts
@@ -1,0 +1,2 @@
+export { default as ErrorBoundary } from './Errorboundary'
+export type { FallbackProps } from './Errorboundary'
