+++ src/components/common/index.ts
@@ -2,7 +2,9 @@ export { Logo } from './Logo'
 export { Spinner } from './Spinner'
 export { Content } from './Content'
 export { PaginationButtons } from './PaginationButtons'
+export { default as ErrorHandlingWrapper } from './ErrorHandlingWrapper'
 
 export * from './PageBreadcrumb'
+export * from './error-boundary'
 
 export { default as MainErrorFallback } from './MainErrorFallback'
