+++ src/lib/axios-instance.ts
@@ -2,7 +2,7 @@ import axios, { AxiosInstance, AxiosRequestConfig } from 'axios'
 
 export const initInstance = (config: AxiosRequestConfig): AxiosInstance => {
   const instance = axios.create({
-    timeout: 5000,
+    timeout: 2000,
     ...config,
     headers: {
       Accept: 'application/json',
