+++ src/lib/query-client.tsx
@@ -4,7 +4,7 @@ import { ReactNode } from 'react'
 
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 
-export const queryClient = new QueryClient({
+const queryClient = new QueryClient({
   defaultOptions: {
     queries: {
       retry: 1,
@@ -15,12 +15,10 @@ export const queryClient = new QueryClient({
   },
 })
 
-export const ReactQueryClientProvider = ({
-  children,
-}: {
-  children: ReactNode
-}) => {
+const ReactQueryClientProvider = ({ children }: { children: ReactNode }) => {
   return (
     <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
   )
 }
+
+export { queryClient, ReactQueryClientProvider }
